Set the theme to either light or dark modes
This demonstrates exporting the HTML in both light and dark modes. This
theme property needs to be set on the exporter itself; it's not
sufficient to just add it to the resources dictionary since it's
overriden by the theme attribute on the exporter class.

Lastly, fixed the issue with setting up and using resources; they need
to be passed to the `from_filename` method of course; not just to the
write_conversion helper function.