Implement algorithm for building document outline
Use markdown-it-py to part the markdown contents of a notebook.

Use a new SectionChildren class that has an `insert_section` method
that takes care of inserting the section, either as a sibling or
recursively calling insert_section into the last child.

Export the toc outline into the resources dictionary as a plain list and
dictionary for portability.